Time
What if time is just an illusion created by the government to keep people away from their own bodies and make them work for hours on end with zero complaints?

Please spread the message about the corruption in our world so people can stop working for the evils in this world. If you don’t believe this it’s because of the corrupt politicians that act like people but are in fact hyper realistic robots who is lead by the corrupt governments around the world. All I ask from you is to spread the message to the rest of the world 🙂